Reconcile with Linux script
This covers a manual review of the windows vs linux scripts for clojure
and updating them as differences came up. Also tried to further clean up
the syntax and conventions to align with PowerShell.

This needs additional testing to be sure the changes didn't break
anything major and also just the full exhaustive sweet of all of the
options.